requirements.txt
fastapi
uvicorn
pydantic
python-multipart
mysql-connector-python
pillow
python-dotenv